Protocol 1001 (1.26.34), .NET 10, and blocks addressed by state

+++ src/MiNET/MiNET/Net/RakNet/ICustomPacketFactory.cs
@@ -34,6 +34,6 @@ namespace MiNET.Net.RakNet
 	/// </summary>
 	public interface ICustomPacketFactory
 	{
-		public Packet Create(byte messageId, ReadOnlyMemory<byte> buffer, string ns);
+		public Packet Create(int messageId, ReadOnlyMemory<byte> buffer, string ns);
 	}
 }
